refactor: change app and  pages structure
