+++ frontend/src/components/pages/Home/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../style.css'
+import './style.css'
 import { Link } from 'react-router-dom';
 import { Breadcrumb, Layout, Menu, theme, Col, Divider, Row, Button, Typography, Space } from 'antd';
