--- frontend/src/routes/AppRouter.tsx
@@ -1,19 +1,0 @@
-import React from "react";
-import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
-import HomePage from "../components/HomePage";
-import LoginPage from "../components/LoginPage";
-import SignUpPage from "../components/SignUpPage";
-
-const AppRouter = () => {
-    return (
-        <Router>
-            <Routes>
-                <Route path="/" Component={HomePage} />
-                <Route path="/login" Component={LoginPage} />
-                <Route path="/signup" Component={SignUpPage} />
-            </Routes>
-        </Router>
-    );
-}
-
-export default AppRouter;
